requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyairtable>=2.0.0
opencv-python>=4.8.0
openai>=1.0.0
textual>=0.40.0

# Development dependencies
ruff>=0.1.0
pytest>=7.0.0
